add sweetalert2

+++ frontend/src/app/tattoos-list/tattoos-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import Swal from 'sweetalert2';
 import { TattoosService } from '../_services/tattoos.service';
 
 @Component({
@@ -32,14 +33,36 @@ export class TattoosListComponent implements OnInit {
 
   deleteTattoo(id: any) { 
     console.log( 'suppression id : ' + id)
-    this.tattooService.delete(id).subscribe((data: any = []) => { 
-      this.tattoos = data 
-      return this.tattoos; 
-    }, (error) => { 
-      console.log('error : ' + error)
+
+    Swal.fire({
+      title: 'Are you sure you want to delete this tattoo ?',
+      showDenyButton: true,
+      confirmButtonText: 'Yep ! delete it',
+      denyButtonText: `No, I changed my mind...`,
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        this.tattooService.delete(id).subscribe((data: any = []) => { 
+          this.tattoos = data 
+          return this.tattoos; 
+        }, (error) => { 
+          console.log('error : ' + error)
+        })
+        setTimeout(() => {
+          console.log('sleep');
+          Swal.fire('Tattoo has been deleted !', '', 'info')
+          // And any other code that should run only after 5s
+        }, 5000);
+        this.reloadPage()
+      } else if (result.isDenied) {
+        Swal.fire('Changes are not saved', '', 'success')
+      } 
     })
-    this.reloadPage()
+
+
   }
+ 
+
 
   reloadPage(): void {
     window.location.reload();
